// src/app/exams/exam/model/exam.ts
export interface Exam {
  examId?: number;
  labAreaId?: number;
  methodologyId?: number;
  examCategoryId?: number;
  sampleTypeId?: number;
  measurementUnitId?: number;
  resultTypeId?: number;
  description?: string;
  shortDescription?: string;
  lisCode?: string;
  count?: string;
  confidential?: string;
  calculated?: string;
  status?: number;
}